+++ app/controllers/postTime.js
@@ -1,6 +1,6 @@
 const express = require("express")
 const fs = require("fs")
-const users = require("../../server/users/user.json")
+const users = require("../users/user.json")
 
 const postTime = (req, res) => {
   const day = new Date()
